Весь функционал работает, тест тестов

+++ src/main/java/rita/controller/resource/ResourceController.java
@@ -1,16 +1,28 @@
 package rita.controller.resource;
 
 
+import io.minio.StatObjectArgs;
+import io.minio.errors.ErrorResponseException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import org.springframework.core.io.InputStreamResource;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import rita.dto.MessageDto;
+import rita.dto.ResourceResponseDto;
+import rita.exeptions.EntityAlreadyExistsException;
+import rita.exeptions.MinioException;
 import rita.service.MinioService;
+import rita.service.NamingService;
 
+import javax.persistence.EntityNotFoundException;
+import java.util.Collections;
 import java.util.List;
 
 @RestController
@@ -20,6 +32,7 @@ import java.util.List;
 public class ResourceController {
 
     private final MinioService minioService;
+    private final NamingService namingService;
 
     @Operation(
             summary = "Скачивает файл или папку"
@@ -27,7 +40,32 @@ public class ResourceController {
     @GetMapping(path = "/download")
     public ResponseEntity<?> downloadResource(@RequestParam("path")
                                               @Parameter(description = "Путь к файлу") String path) {
-        return minioService.downloadResource(path);
+
+        if (path == null || path.isEmpty()) {
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body(new MessageDto("Невалидный или отсутствующий путь"));
+        }
+        InputStreamResource resource;
+        try {
+            resource = minioService.downloadResource(path);
+        } catch (EntityNotFoundException e) {
+            return ResponseEntity
+                    .status(HttpStatus.NOT_FOUND)
+                    .body(new MessageDto("Ресурс не найден"));
+        } catch (MinioException e) {
+            return ResponseEntity
+                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(new MessageDto("Неизвестная ошибка при скачивании файла"));
+        }
+        String filename = namingService.getNameFromPath(path);
+        if (!filename.endsWith(".zip")) {
+            filename += ".zip";
+        }
+        return ResponseEntity.ok()
+                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + filename + "\"")
+                .contentType(MediaType.APPLICATION_OCTET_STREAM)
+                .body(resource);
     }
 
     @Operation(
@@ -36,10 +74,23 @@ public class ResourceController {
     @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<?> uploadResource(@RequestParam("object")
                                             @Parameter(description = "Список файлов в формате MultipartFile")
-                                                List<MultipartFile> file,
+                                            List<MultipartFile> file,
                                             @Parameter(description = "Путь к файлу")
                                             @RequestParam("path") String path) {
-        return minioService.uploadFile(file, path);
+
+        List<ResourceResponseDto> files;
+
+        try {
+            files = minioService.uploadFile(file, path);
+        } catch (EntityAlreadyExistsException e) {
+            return ResponseEntity.status(HttpStatus.CONFLICT)
+                    .body(new MessageDto("Файл с таким именем уже существует"));
+        } catch (MinioException e) {
+            return ResponseEntity
+                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(new MessageDto("Неизвестная ошибка при загрузке файла"));
+        }
+        return ResponseEntity.status(HttpStatus.CREATED).body(files);
     }
 
     @Operation(
@@ -48,7 +99,21 @@ public class ResourceController {
     @GetMapping
     public ResponseEntity<?> getInfo(@RequestParam("path")
                                      @Parameter(description = "Путь к файлу") String path) {
-        return minioService.getInfo(path);
+        if (path == null || path.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(new MessageDto("Невалидный или отсутствующий путь"));
+        }
+        ResourceResponseDto dto;
+        try {
+            dto = minioService.getInfo(path);
+        } catch (EntityNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(new MessageDto("Ресурс не найден"));
+        } catch (MinioException e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(new MessageDto("Неизвестная ошибка: " + e.getMessage()));
+        }
+        return ResponseEntity.ok(dto);
     }
 
     @Operation(
@@ -56,8 +121,21 @@ public class ResourceController {
     )
     @DeleteMapping
     public ResponseEntity<?> deleteResource(@RequestParam("path")
-                                     @Parameter(description = "Путь к файлу") String path) {
-       return minioService.deleteResource(path);
+                                            @Parameter(description = "Путь к файлу") String path) {
+        if (path == null || path.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(new MessageDto("Невалидный или отсутствующий путь"));
+        }
+        try {
+            minioService.deleteResource(path);
+        } catch (EntityNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(new MessageDto("Ресурс не найден"));
+        } catch (MinioException e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(new MessageDto("Ошибка при удалении ресурсов: " + e.getMessage()));
+        }
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
     @Operation(
@@ -68,7 +146,22 @@ public class ResourceController {
                                           @Parameter(description = "Откуда перемещается файл") String from,
                                           @Parameter(description = "Куда перемещается файл")
                                           @RequestParam("to") String to) {
-        return minioService.moveOrRenameResource(from, to);
+        ResourceResponseDto dto;
+
+        if (from.equals(to)) {
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
+        try {
+            dto = minioService.moveOrRenameResource(from, to);
+        } catch (MinioException e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(new MessageDto("Ошибка при работе с ресурсами: " + e.getMessage()));
+
+        } catch (EntityAlreadyExistsException e) {
+        return ResponseEntity.status(HttpStatus.CONFLICT)
+                .body(new MessageDto("Файл с таким именем уже существует"));
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(dto);
     }
 
     @Operation(
@@ -78,6 +171,15 @@ public class ResourceController {
     public ResponseEntity<?> searchResource(@RequestParam("query")
                                             @Parameter(description = "Строка, которую вводит пользователь в поиск")
                                             String query) {
-        return minioService.searchResource(query);
+        List<ResourceResponseDto> files;
+
+        try {
+            files = minioService.searchResource(query);
+
+        } catch (MinioException e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(new MessageDto("Ошибка при поиске ресурсов: " + e.getMessage()));
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(files);
     }
 }
